+++ lspattern/blocks/pipes/measure.py
@@ -7,6 +7,7 @@ from graphix_zx.common import Axis, AxisMeasBasis, Sign
 from graphix_zx.graphstate import GraphState
 
 from lspattern.blocks.pipes.base import RHGPipe, RHGPipeSkeleton
+from lspattern.consts import PatchType
 from lspattern.mytype import NodeIdLocal, PatchCoordGlobal3D, PhysCoordGlobal3D, PhysCoordLocal2D, SpatialEdgeSpec
 from lspattern.tiling.template import RotatedPlanarPipetemplate
 from lspattern.utils import get_direction
@@ -44,7 +45,7 @@ class _MeasurePipeBase(RHGPipe):
         if patch_coord is not None and self.source is not None and self.sink is not None:
             source_2d = (self.source[0], self.source[1])
             sink_2d = (self.sink[0], self.sink[1])
-            idx_map = self.template.get_data_indices(source_2d, patch_type="pipe", sink_patch=sink_2d)
+            idx_map = self.template.get_data_indices(source_2d, patch_type=PatchType.PIPE, sink_patch=sink_2d)
         else:
             idx_map = self.template.get_data_indices()
         self.in_ports = set(idx_map.values())